// src/inmankist/enneagram/types.ts
export enum EnneagramType {
  Type1 = "type1", // The Reformer
  Type2 = "type2", // The Helper
  Type3 = "type3", // The Achiever
  Type4 = "type4", // The Individualist
  Type5 = "type5", // The Investigator
  Type6 = "type6", // The Loyalist
  Type7 = "type7", // The Enthusiast
  Type8 = "type8", // The Challenger
  Type9 = "type9", // The Peacemaker
}